sm-backend/stock_monitor/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
dj-rest-auth==6.0.0
Django==5.0.7
django-allauth==64.0.0
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.2
environ==1.0
frozendict==2.4.4
html5lib==1.1
idna==3.7
lxml==5.2.2
multitasking==0.0.11
mysql-connector-python==9.0.0
mysqlclient==2.2.4
numpy==2.0.1
pandas==2.2.2
peewee==3.17.6
platformdirs==4.2.2
prometheus_client==0.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
six==1.16.0
soupsieve==2.5
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
webencodings==0.5.1
yfinance==0.2.41
